chore: Move dispatcher tests to less inconvenient namespace
