+++ src/Uno.UI.RuntimeTests/Tests/Windows_System/Given_DispatcherQueue.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using Windows.System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Uno.UI.RuntimeTests.Tests.System
+namespace Uno.UI.RuntimeTests.Tests.Windows_System
 {
 	[TestClass]
 	public class Given_DispatcherQueue
